requirements.txt
Appium-Python-Client==3.1.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
exceptiongroup==1.1.3
h11==0.14.0
html5lib==1.1
HTMLParser==0.0.2
idna==3.4
numpy==1.26.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
webencodings==0.5.1
wsproto==1.2.0